backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
spacy
Pillow
pytesseract
